EICNET-1225: Move code from eic_group_statistics to eic_media_statistics module.
