+++ lib/modules/eic_media/modules/eic_media_statistics/eic_media_statistics.module
@@ -6,6 +6,7 @@
  */
 
 use Drupal\Core\Cache\Cache;
+use Drupal\Core\Entity\Display\EntityViewDisplayInterface;
 use Drupal\Core\Entity\EntityInterface;
 
 /**
@@ -46,4 +47,26 @@ function eic_media_statistics_file_predelete(EntityInterface $entity) {
   // Clean up statistics table when file is deleted.
   $id = $entity->id();
   return \Drupal::service('eic_media_statistics.storage.file')->deleteViews($id);
+}
+
+/**
+ * Implements hook_ENTITY_TYPE_view().
+ */
+function eic_media_statistics_node_view(array &$build, EntityInterface $entity, EntityViewDisplayInterface $display, $view_mode) {
+  switch ($entity->bundle()) {
+    case 'document':
+    case 'gallery':
+      // We need to make sure the node belongs to a group content.
+      $group_contents = GroupContent::loadByEntity($entity);
+      if (!$group_contents) {
+        return;
+      }
+
+      $group_content = reset($group_contents);
+
+      \Drupal::classResolver(EntityOperations::class)
+        ->groupContentNodeView($build, $entity, $display, $view_mode, $group_content);
+      break;
+
+  }
 }
