+++ lib/modules/eic_statistics/modules/eic_group_statistics/eic_group_statistics.module
@@ -5,7 +5,6 @@
  * Primary module hooks for EIC Group Statistics module.
  */
 
-use Drupal\Core\Entity\Display\EntityViewDisplayInterface;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\eic_group_statistics\GroupStatisticTypes;
 use Drupal\eic_group_statistics\Hooks\EntityOperations;
@@ -147,28 +146,6 @@ function eic_group_statistics_node_update(EntityInterface $entity) {
 
       \Drupal::classResolver(EntityOperations::class)
         ->groupContentNodeUpdate($entity, $group_content);
-      break;
-
-  }
-}
-
-/**
- * Implements hook_ENTITY_TYPE_view().
- */
-function eic_group_statistics_node_view(array &$build, EntityInterface $entity, EntityViewDisplayInterface $display, $view_mode) {
-  switch ($entity->bundle()) {
-    case 'document':
-    case 'gallery':
-      // We need to make sure the node belongs to a group content.
-      $group_contents = GroupContent::loadByEntity($entity);
-      if (!$group_contents) {
-        return;
-      }
-
-      $group_content = reset($group_contents);
-
-      \Drupal::classResolver(EntityOperations::class)
-        ->groupContentNodeView($build, $entity, $display, $view_mode, $group_content);
       break;
 
   }
